+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     - Remove superfluous autoinstall feature that was superseded by the statically provided web service
 - Finalize Moodle privacy API provider
 - Adapt web service unit tests to latest activity archiving task access token invalidation behavior
+- Create unit tests for activity archiving driver base class
 
 
 ## Version 0.3.0 (2025101300)